Split XDG to modules
