+++ src/wayland/shell/xdg/handlers.rs
@@ -1,0 +1,15 @@
+use super::{
+    InnerState, PopupConfigure, PositionerState, ShellClient, ShellClientData, SurfaceCachedState,
+    ToplevelConfigure, XdgPopupSurfaceRoleAttributes, XdgRequest, XdgShellDispatch, XdgShellHandler,
+    XdgToplevelSurfaceRoleAttributes,
+};
+
+mod wm_base;
+pub use wm_base::XdgWmBaseUserData;
+
+mod positioner;
+pub use positioner::XdgPositionerUserData;
+
+mod surface;
+pub(super) use surface::{get_parent, send_popup_configure, send_toplevel_configure, set_parent};
+pub use surface::{XdgShellSurfaceUserData, XdgSurfaceUserData};
